Accidentally started dev on old alpha folder and needed to merge changes. Fixed auto layout errors. Added toolbar and tab bar icons from PaintKit stylekit. Moved Rules popover into appearance popover/tabbar. Changed backgrounds to iOS 7 translucent style.

+++ FractalScape/GradientView.m
@@ -63,7 +63,7 @@
         
         _colors = [colors copy];
         
-        int componentCount = colors.count * COLOR_COMPONENT_COUNT;
+        unsigned long componentCount = colors.count * COLOR_COMPONENT_COUNT;
         CGFloat gradientColors[componentCount];
         int counter = 0;
         for (UIColor *color in colors) {
